refactor(filters): PreservedParams y Persistable compartidos; el form simple re-emite el query de su url
Extrae a Bali::Filters::PreservedParams la semántica de preservación de
params que Filters::Component ya tenía (query de la url: menos
EXCLUDED_PARAMS, hash explícito gana en colisión de key) y la adopta
SimpleFilters: un host que pasa url: con params propios (un scope como
status=historico) los perdía en cada submit GET del form simple — el
link Limpiar los conservaba desde beta.6, el submit no.

También extrae el trío storage_id / persist_enabled? /
persistence_toggle? al concern Bali::Filters::Persistable, que ambos
components copiaban.

Refs #725
